backend/requirements.txt
fastapi
uvicorn[standard]
PyYAML==6.0.1
ansible
pydantic
httpx
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
requests
